+++ build-doc.sh
@@ -30,10 +30,10 @@ echo "\`\`\`" >> README.md
 ${CMD} >> README.md
 echo "\`\`\`" >> README.md
 echo "" >> README.md
-echo "## Sample" >> README.md
+echo "## Analyse" >> README.md
 echo "" >> README.md
 echo "\`\`\`" >> README.md
-${CMD} sample >> README.md
+${CMD} analyse >> README.md
 echo "\`\`\`" >> README.md
 echo "" >> README.md
 echo "## Compare" >> README.md
@@ -42,20 +42,35 @@ echo "\`\`\`" >> README.md
 ${CMD} compare >> README.md
 echo "\`\`\`" >> README.md
 echo "" >> README.md
-echo "## Reindex" >> README.md
+echo "## Degrade" >> README.md
 echo "" >> README.md
 echo "\`\`\`" >> README.md
-${CMD} reindex >> README.md
+${CMD} degrade >> README.md
+echo "\`\`\`" >> README.md
+echo "## Dequantize" >> README.md
+echo "" >> README.md
+echo "\`\`\`" >> README.md
+${CMD} dequantize >> README.md
+echo "\`\`\`" >> README.md
+echo "## Normals" >> README.md
+echo "" >> README.md
+echo "\`\`\`" >> README.md
+${CMD} normals >> README.md
 echo "\`\`\`" >> README.md
 echo "## Quantize" >> README.md
 echo "" >> README.md
 echo "\`\`\`" >> README.md
 ${CMD} quantize >> README.md
 echo "\`\`\`" >> README.md
-echo "## Degrade" >> README.md
+echo "## Reindex" >> README.md
 echo "" >> README.md
 echo "\`\`\`" >> README.md
-${CMD} degrade >> README.md
+${CMD} reindex >> README.md
+echo "\`\`\`" >> README.md
+echo "## Sample" >> README.md
+echo "" >> README.md
+echo "\`\`\`" >> README.md
+${CMD} sample >> README.md
 echo "\`\`\`" >> README.md
 echo "## Sequence" >> README.md
 echo "" >> README.md