+++ test.sh
@@ -1,5 +1,4 @@
 #!/bin/bash
 
 cd test
-./test-all.sh
-
+./test-all.sh $@