+++ test/config.sh
@@ -10,7 +10,7 @@ DATA=./data
 EXTDATA=./extdata
 # path to test output references
 REFS=./refs
-# make a tmp folder to store processing ouotputs
+# make a tmp folder to store processing outputs
 TMP=./tmp
 mkdir -p $TMP
 
@@ -27,15 +27,17 @@ fi
 function cmpOsLog {
 	if [ "$(uname)" == "Linux" ]; 
 	then
-		cmp ${REFS}/${1}_linux.txt ${TMP}/${1}.txt
+		# first apply sed to change command name
+		# '-' in diff part means use std input (from sed)
+		sed 's/mm.exe/mm/' ${REFS}/${1}.txt | diff -a - ${TMP}/${1}.txt
 	else
-		cmp ${REFS}/${1}.txt ${TMP}/${1}.txt
+		diff -a ${REFS}/${1}.txt ${TMP}/${1}.txt
 	fi
 }
 
 # test if $1 file contains $3 times the $2 string
 function fileHasString {
 	if [ $(grep -c "$2" $1) -ne $3 ]; then 
-		echo "Error: $1 expected grep return code = $3, got $?"
+		echo "Error: expected $3 time(s) the string \"$2\" in file $1"
 	fi
 }