+++ test/test-all.sh
@@ -7,17 +7,20 @@
 
 tests=(
 "test-help"
-"test-quantize"
-"test-reindex"
-"test-degrade"
-"test-sample-face"
-"test-sample-sdiv"
-"test-sample-grid"
-"test-sample-map"
+"test-analyse"
 "test-compare-eq" 
 "test-compare-pcc"
 "test-compare-pcqm"
 "test-composed"
+"test-degrade"
+"test-normals"
+"test-quantize"
+"test-reindex"
+"test-sample-ediv"
+"test-sample-face"
+"test-sample-grid"
+"test-sample-map"
+"test-sample-sdiv"
 )
 
 for test in ${tests[@]}; do