+++ test/test-compare-pcc.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+# attention: using only 1 digit after coma for PSNR results comparison due to imprecisions in linux/windows
+
 source config.sh
 
 # compare pcc
@@ -12,7 +14,7 @@ fileHasString ${TMP}/${OUT}.txt "mseF,PSNR (p2plane): inf" 1
 # no map, no color
 OUT=compare_pcc_sphere_qp8
 $CMD compare --mode pcc --inputModelA ${DATA}/sphere.obj --inputModelB ${DATA}/sphere_qp8.obj > ${TMP}/${OUT}.txt 2>&1
-fileHasString ${TMP}/${OUT}.txt "mseF,PSNR (p2plane): 66.4296" 1
+fileHasString ${TMP}/${OUT}.txt "mseF,PSNR (p2plane): 66.4" 1
 
 # external dataset
 if [ "$1" == "ext" ]; 
@@ -40,12 +42,21 @@ then
 	fileHasString ${TMP}/${OUT}.txt "mseF,PSNR (p2plane): inf" 1
 	fileHasString ${TMP}/${OUT}.txt "c\[0\],PSNRF         : inf" 1
 
-	OUT=compare_pcc_basket_qp8
+	OUT=compare_pcc_basket_qp8_orig
 	$CMD compare --mode pcc \
 		--inputModelA ${EXTDATA}/basketball_player_00000001.obj --inputMapA  ${EXTDATA}/basketball_player_00000001.png \
-		--inputModelB ${DATA}/basketball_player_00000001_qp8_orig.obj --inputMapB  ${EXTDATA}/basketball_player_00000001.png  > ${TMP}/${OUT}.txt 2>&1
-	fileHasString ${TMP}/${OUT}.txt "mseF,PSNR (p2plane): 67.6033" 1
-	fileHasString ${TMP}/${OUT}.txt "c\[0\],PSNRF         : 31.9359" 1
+		--inputModelB ${DATA}/basketball_player_00000001_qp8_orig.obj --inputMapB  ${EXTDATA}/basketball_player_00000001.png \
+		> ${TMP}/${OUT}.txt 2>&1
+	fileHasString ${TMP}/${OUT}.txt "mseF,PSNR (p2plane): 67.6" 1
+	fileHasString ${TMP}/${OUT}.txt "c\[0\],PSNRF         : 31.9" 1
+
+	OUT=compare_pcc_basket_qp8_hole
+	$CMD compare --mode pcc \
+		--inputModelA ${EXTDATA}/basketball_player_00000001.obj --inputMapA  ${EXTDATA}/basketball_player_00000001.png \
+		--inputModelB ${DATA}/basketball_player_00000001_qp8_hole.obj --inputMapB  ${EXTDATA}/basketball_player_00000001.png \
+		> ${TMP}/${OUT}.txt 2>&1
+	fileHasString ${TMP}/${OUT}.txt "mseF,PSNR (p2plane): 67.6" 1
+	fileHasString ${TMP}/${OUT}.txt "c\[0\],PSNRF         : 31.9" 1
 
 	OUT=compare_pcc_basket_qp16_self
 	$CMD compare --mode pcc \
@@ -58,14 +69,14 @@ then
 	$CMD compare --mode pcc \
 		--inputModelA ${EXTDATA}/basketball_player_00000001.obj --inputMapA  ${EXTDATA}/basketball_player_00000001.png \
 		--inputModelB ${DATA}/basketball_player_00000001_qp16.obj --inputMapB  ${EXTDATA}/basketball_player_00000001.png  > ${TMP}/${OUT}.txt 2>&1
-	fileHasString ${TMP}/${OUT}.txt "mseF,PSNR (p2plane): 115.398" 1
-	fileHasString ${TMP}/${OUT}.txt "c\[0\],PSNRF         : 38.1793" 1	
+	fileHasString ${TMP}/${OUT}.txt "mseF,PSNR (p2plane): 115.3" 1
+	fileHasString ${TMP}/${OUT}.txt "c\[0\],PSNRF         : 38.1" 1	
 
 	OUT=compare_pcc_basket_qp16_nomap
 	$CMD compare --mode pcc \
 		--inputModelA ${EXTDATA}/basketball_player_00000001.obj  \
 		--inputModelB ${DATA}/basketball_player_00000001_qp16.obj  > ${TMP}/${OUT}.txt 2>&1
-	fileHasString ${TMP}/${OUT}.txt "mseF,PSNR (p2plane): 115.398" 1
+	fileHasString ${TMP}/${OUT}.txt "mseF,PSNR (p2plane): 115.3" 1
 
 	# test sequence mode with self
 	OUT=compare_pcc_longdress_1K_seq
@@ -79,3 +90,5 @@ then
 	grep -iF "error" ${TMP}/${OUT}.txt
 
 fi
+
+# EOF